# server/openapi/aro-sync-v2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aro Sync Protocol",
    "version": "2.0.0"
  },
  "servers": [{"url": "https://{hub}:4848", "variables": {"hub": {"default": "aro.local"}}}],
  "paths": {
...
    "/v1/negotiate": {"post": {"summary": "Negotiate a protocol version", "responses": {"200": {"description": "Selected version"}, "426": {"description": "Incompatible"}}}},
...
    "/v1/blobs/{hash}/status": {"get": {"summary": "Inspect committed and partial blob state", "parameters": [{"name": "hash", "in": "path", "required": true}], "responses": {"200": {"description": "Blob state"}}}},
    "/v1/blobs/{hash}": {
      "get": {"summary": "Download a blob, optionally by byte range", "parameters": [{"name": "hash", "in": "path", "required": true}], "responses": {"200": {"description": "Blob"}, "206": {"description": "Blob range"}}},
      "put": {"summary": "Append a resumable upload chunk", "parameters": [{"name": "hash", "in": "path", "required": true}, {"name": "X-Aro-Offset", "in": "header", "required": true}], "responses": {"200": {"description": "Upload state"}}}
    },
...
  }
}